change chatter to forum
