+++ frontend-cosmic-chatter/src/app/profile/page.tsx
@@ -18,7 +18,7 @@ const page = () => {
             </h4>
             <div className='flex justify-start my-3'>
               <img
-                className='w-full m-3  h-10 w-10 rounded-full ring-2 ring-white'
+                className=' m-3  h-10 w-10 rounded-full ring-2 ring-white'
                 src='https://images.unsplash.com/photo-1491528323818-fdd1faba62cc?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80'
                 alt='avatar image'
               ></img>
